# pkg/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "invenlore.io OpenAPI Swagger",
    "version": "latest"
  },
  "tags": [
    {
      "name": "IdentityService"
    }
  ],
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/users": {
      "get": {
        "summary": "List users",
        "description": "List all users",
        "operationId": "IdentityService_ListUsers",
        "responses": {
          "200": {
            "description": "(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/identityUser"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of identityListUsersResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Users"
        ]
      },
      "post": {
        "summary": "Add user",
        "description": "Add user",
        "operationId": "IdentityService_AddUser",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/identityAddUserRequest"
            }
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "summary": "Get user",
        "description": "Get user by ID",
        "operationId": "IdentityService_GetUser",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/identityUser"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "summary": "Delete user",
        "description": "Delete user by ID",
        "operationId": "IdentityService_DeleteUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/identityDeleteUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Users"
        ]
      }
    }
  },
  "definitions": {
    "identityAddUserRequest": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/identityUser"
        }
      },
      "title": "Add User request"
    },
    "identityAddUserResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      },
      "title": "Add User response"
    },
    "identityDeleteUserResponse": {
      "type": "object",
      "title": "Delete User response"
    },
    "identityGetUserResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/identityUser"
        }
      },
      "title": "Get User response"
    },
    "identityHealthResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        }
      },
      "title": "HealthCheck response"
    },
    "identityListUsersResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/identityUser"
        }
      },
      "title": "List Users response"
    },
    "identityUser": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      },
      "title": "User"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  },
  "externalDocs": {
    "description": "gRPC-gateway invenlore proto repository",
    "url": "https://github.com/invenlore/proto"
  }
}